patchpilot/backend/requirements.txt
fastapi
uvicorn
celery
redis
motor
pymongo
openai
PyGithub
gitpython
python-dotenv
